+++ packages/eas-cli/src/rollout/actions/EditRollout.ts
@@ -51,8 +51,14 @@ export class EditRollout implements EASUpdateAction<UpdateChannelBasicInfoFragme
     const channelObject = await this.getChannelObjectAsync(ctx);
     const rollout = getRollout(channelObject);
     const { rolledOutBranch, defaultBranch } = rollout;
-    const promptMessage = `What percent of users should be directed to the ${rolledOutBranch.name} branch ?`;
+    const promptMessage = `What percent of users should be sent to the ${rolledOutBranch.name} branch ?`;
     const percent = this.options.percent ?? (await promptForRolloutPercentAsync({ promptMessage }));
+
+    if (percent === 0 || percent === 100) {
+      Log.warn(
+        `Editing the percent to ${percent} will not end the rollout. You'll need to end the rollout from the main menu.`
+      );
+    }
 
     const oldBranchMapping = getRolloutBranchMapping(channelObject.branchMapping);
     const newBranchMapping = editRolloutBranchMapping(oldBranchMapping, percent);